requirements.txt
Flask==3.0.0
Authlib==1.3.0
Flask-Session==0.5.0
PyJWT==2.8.0
requests==2.31.0
cryptography==41.0.7
gunicorn==21.2.0
